agent-automation/mcp-server/requirements.txt
mcp>=0.9.0
PyYAML>=6.0

